[ADVAPP-1669]: Relocate action center for students and for prospects to the students and prospects portion of the Report Library
